Fix VS 2019 IntelliSense warnings

+++ XRVessels/XR5Vanguard/XR5Vanguard/XR5Globals.cpp
@@ -234,7 +234,7 @@ const double HIDDEN_ELEVATOR_TRIM_STATE = -0.598760;
 // NOTE: we assume a LOX mass here of a 14-day loadout @ 'realistic': 25.6 kg per passenger
 // NOTE: max landing mass does NOT change with cheatcodes!
 // we use an "average" passenger mass of 68 kg for landing limit calculations
-const double FULLY_LOADED_MASS = (EMPTY_MASS + (68.0 * MAX_PASSENGERS) + TANK1_CAPACITY + TANK2_CAPACITY + RCS_FUEL_CAPACITY + APU_FUEL_CAPACITY + (25.6 * MAX_PASSENGERS));
+const double FULLY_LOADED_MASS = (EMPTY_MASS + (68.0 * static_cast<double>(MAX_PASSENGERS)) + TANK1_CAPACITY + TANK2_CAPACITY + RCS_FUEL_CAPACITY + APU_FUEL_CAPACITY + (25.6 * MAX_PASSENGERS));
 
 // rate at which main fuel is dumped and loaded in kg/sec
 const double FUEL_DUMP_RATE = 85 * 22.2;
